refactor(webapp): extract OTLP transform into a worker-safe module
Move the pure OTLP decode/convert/attribute helpers out of otlpExporter.server.ts
into otlpTransform.server.ts, which has no server-singleton imports (env, clickhouse
factory, event repository, prisma) so it can run inside a worker thread. The env-derived
event-store default is now passed as a parameter. Behavior is identical; typecheck green.

Prep for the OTLP ingest worker-pool spike.
